+++ paddle/fluid/pir/dialect/op_generator/op_gen.py
@@ -88,7 +88,7 @@ H_FILE_TEMPLATE = """// This file is generated by "paddle/fluid/pir/dialect/op_g
 
 #include "paddle/pir/include/core/builder.h"
 #include "paddle/pir/include/core/operation_utils.h"
-#include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape.h"
+#include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/infer_symbolic_shape.h"
 #include "paddle/pir/include/core/op_base.h"
 #include "paddle/pir/include/core/op_trait.h"
 #include "paddle/fluid/pir/dialect/operator/utils/utils.h"
