+++ paddle/fluid/pir/dialect/operator/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(ir)
 add_subdirectory(interface)
+add_subdirectory(interface/infer_symbolic_shape)
 add_subdirectory(trait)
 add_subdirectory(utils)
 add_subdirectory(transforms)
