+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/infer_sym_utils.h
@@ -1,0 +1,42 @@
+// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "paddle/pir/include/dialect/shape/utils/shape_analysis.h"
+
+namespace paddle::dialect::details {
+template <typename T>
+struct AttributeTrait;
+
+template <>
+struct AttributeTrait<std::int64_t> {
+  using value_type = ::pir::Int64Attribute;
+};
+
+template <>
+struct AttributeTrait<int> {
+  using value_type = ::pir::Int32Attribute;
+};
+
+template <typename T = int64_t>
+std::vector<T> GetVectorAttr(const ::pir::Operation *op,
+                             const std::string &name);
+
+bool ReduceInferDim(pir::Operation *op,
+                    pir::ShapeConstraintIRAnalysis *shape_analysis,
+                    const std::vector<int64_t> &axis,
+                    bool keep_dim,
+                    bool reduce_all);
+}  // namespace paddle::dialect::details
